Let `dataset_id` by a UUID5 (deterministic from raw ID)
This is needed for the catalog tooling/schema-compliance.

To make this flexible, introduce a formating mechanism that can yield
a raw ID string as input for the UUID5 generation. It can be provided
via a new/simple config option. Example:

```
bin/tubby2catalog \
  -t dataset \
  -c 'dataset_id_fmt={additional_display[sfb1451][project]}{dataset_id}'\
  data/dataset_metadata.tsv
```

This configures the UUID5 got be generated from the concatenation of the
'sfb1451'-'project' custom property, plus the 'dataset_id' field.

To make this work, this uses an intermediate metadata representation
that is comprised of mostly dicts, and supports direct addressing of
most properties.